// src/api/generated/model/getActionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */
import type { ActionStep } from './actionStep';
import type { GetActionResponseTriggerConfig } from './getActionResponseTriggerConfig';
import type { Time } from './time';

export interface GetActionResponse {
  active?: boolean;
  created_at?: Time;
  database_id?: string;
  description?: string;
  failure_count?: number;
  id?: string;
  last_error?: string;
  last_run_at?: Time;
  name?: string;
  run_count?: number;
  space_id?: string;
  space_name?: string;
  steps?: ActionStep[];
  success_count?: number;
  trigger_config?: GetActionResponseTriggerConfig;
  trigger_type?: string;
  updated_at?: Time;
}