// src/api/generated/model/createActionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */
import type { ActionStep } from './actionStep';
import type { CreateActionRequestTriggerConfig } from './createActionRequestTriggerConfig';

export interface CreateActionRequest {
  database_id?: string;
  description?: string;
  name?: string;
  space_id?: string;
  steps?: ActionStep[];
  trigger_config?: CreateActionRequestTriggerConfig;
  trigger_type?: string;
}