// src/api/generated/model/updateActionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */
import type { ActionStep } from './actionStep';
import type { UpdateActionRequestTriggerConfig } from './updateActionRequestTriggerConfig';

export interface UpdateActionRequest {
  active?: boolean;
  description?: string;
  name?: string;
  steps?: ActionStep[];
  trigger_config?: UpdateActionRequestTriggerConfig;
  trigger_type?: string;
}